+++ tests/test_l.rs
@@ -1,136 +1,45 @@
-macro_rules! help_msg {
-    () => {
-        concat!(
-            version_msg!(),
-            "\n",
-            indoc::indoc!(r#"
-            Usage:
-              aki-json-pick [options]
+#[macro_use]
+mod helper;
 
-            The json pick out command.
+#[macro_use]
+mod helper_l;
 
-            Options:
-                  --color <when>        json colored output.
-              -s, --select <selector>   pick out json value by <selector>.
-              -p, --pretty              pretty output.
-              -r, --raw-output          raw string output without JSON double-quote.
-
-              -H, --help        display this help and exit
-              -V, --version     display version information and exit
-              -X <x-options>    x options. try -X help
-
-            Option Parameters:
-              <when>        'always', 'never', or 'auto'
-              <selector>    json selector
-
-            Examples:
-              pick out some.property value:
-                echo -e '{ "some": { "property": "yay!" } }' | aki-json-pick -s '"some"."property"'
-            "#),
-            "\n",
-        )
-    };
-}
-
-macro_rules! try_help_msg {
-    () => {
-        "Try --help for help.\n"
-    };
-}
-
-macro_rules! program_name {
-    () => {
-        "aki-json-pick"
-    };
-}
-
-macro_rules! version_msg {
-    () => {
-        concat!(program_name!(), " ", env!("CARGO_PKG_VERSION"), "\n")
-    };
-}
-
-/*
-macro_rules! fixture_text10k {
-    () => {
-        "fixtures/text10k.txt"
-    };
-}
-*/
-
-macro_rules! do_execute {
-    ($args:expr) => {
-        do_execute!($args, "")
-    };
-    ($args:expr, $sin:expr) => {{
-        let sioe = RunnelIoe::new(
-            Box::new(StringIn::with_str($sin)),
-            #[allow(clippy::box_default)]
-            Box::new(StringOut::default()),
-            #[allow(clippy::box_default)]
-            Box::new(StringErr::default()),
-        );
-        let program = env!("CARGO_PKG_NAME");
-        let r = execute(&sioe, &program, $args);
-        match r {
-            Ok(_) => {}
-            Err(ref err) => {
-                let _ = sioe
-                    .pg_err()
-                    .lock()
-                    .write_fmt(format_args!("{}: {}\n", program, err));
-            }
-        };
-        (r, sioe)
-    }};
-}
-
-macro_rules! buff {
-    ($sioe:expr, serr) => {
-        $sioe.pg_err().lock().buffer_to_string()
-    };
-    ($sioe:expr, sout) => {
-        $sioe.pg_out().lock().buffer_to_string()
-    };
-}
-
-mod test_s0 {
+mod test_0_l {
     use libaki_json_pick::*;
     use runnel::medium::stringio::{StringErr, StringIn, StringOut};
     use runnel::RunnelIoe;
-    use std::io::Write;
     //
     #[test]
     fn test_help() {
-        let (r, sioe) = do_execute!(&["-H"]);
+        let (r, sioe) = do_execute!(["-H"]);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), help_msg!());
         assert!(r.is_ok());
     }
     #[test]
     fn test_help_long() {
-        let (r, sioe) = do_execute!(&["--help"]);
+        let (r, sioe) = do_execute!(["--help"]);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), help_msg!());
         assert!(r.is_ok());
     }
     #[test]
     fn test_version() {
-        let (r, sioe) = do_execute!(&["-V"]);
+        let (r, sioe) = do_execute!(["-V"]);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), version_msg!());
         assert!(r.is_ok());
     }
     #[test]
     fn test_version_long() {
-        let (r, sioe) = do_execute!(&["--version"]);
+        let (r, sioe) = do_execute!(["--version"]);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), version_msg!());
         assert!(r.is_ok());
     }
     #[test]
     fn test_non_option() {
-        let (r, sioe) = do_execute!(&[""]);
+        let (r, sioe) = do_execute!([""]);
         #[rustfmt::skip]
         assert_eq!(
             buff!(sioe, serr),
@@ -138,6 +47,73 @@ mod test_s0 {
                 program_name!(), ": ",
                 "Missing option: s\n",
                 "Unexpected argument: \n",
+                try_help_msg!()
+            )
+        );
+        assert_eq!(buff!(sioe, sout), "");
+        assert!(r.is_err());
+    }
+}
+
+mod test_0_x_options_l {
+    use libaki_json_pick::*;
+    use runnel::medium::stringio::*;
+    use runnel::*;
+    //
+    #[test]
+    fn test_x_option() {
+        let (r, sioe) = do_execute!(["-X"]);
+        #[rustfmt::skip]
+        assert_eq!(
+            buff!(sioe, serr),
+            concat!(
+                program_name!(),
+                ": ",
+                "Missing option argument: X\n",
+                "Missing option: s\n",
+                try_help_msg!()
+            )
+        );
+        assert_eq!(buff!(sioe, sout), "");
+        assert!(r.is_err());
+    }
+    //
+    #[test]
+    fn test_x_option_help() {
+        let (r, sioe) = do_execute!(["-X", "help"]);
+        assert_eq!(buff!(sioe, serr), "");
+        assert_eq!(buff!(sioe, sout), x_help_msg!());
+        assert!(r.is_ok());
+    }
+    //
+    #[test]
+    fn test_x_option_rust_version_info() {
+        let (r, sioe) = do_execute!(["-X", "rust-version-info"]);
+        assert_eq!(buff!(sioe, serr), "");
+        assert!(buff!(sioe, sout).contains("rustc"));
+        assert!(r.is_ok());
+    }
+    //
+    #[test]
+    fn test_multiple_x_options() {
+        let (r, sioe) = do_execute!(["-X", "help", "-X", "rust-version-info"]);
+        assert_eq!(buff!(sioe, serr), "");
+        // The first one should be executed and the program should exit.
+        assert!(buff!(sioe, sout).contains("Options:"));
+        assert!(!buff!(sioe, sout).contains("rustc"));
+        assert!(r.is_ok());
+    }
+    #[test]
+    fn test_x_option_invalid() {
+        let (r, sioe) = do_execute!(["-X", "red"]);
+        #[rustfmt::skip]
+        assert_eq!(
+            buff!(sioe, serr),
+            concat!(
+                program_name!(),
+                ": ",
+                "Invalid option argument: X: can not parse 'red'\n",
+                "Missing option: s\n",
                 try_help_msg!()
             )
         );
@@ -158,16 +134,15 @@ const IN_DAT_09: &str = include_str!("../fixtures/09.json");
 const IN_DAT_10: &str = include_str!("../fixtures/10.json");
 const IN_DAT_11: &str = include_str!("../fixtures/11.json");
 
-mod test_s1 {
+mod test_1_l {
     use libaki_json_pick::*;
     use runnel::medium::stringio::{StringErr, StringIn, StringOut};
     use runnel::RunnelIoe;
-    use std::io::Write;
     //
     #[test]
     fn test_t01() {
         // Root selection
-        let (r, sioe) = do_execute!(&["-s", "."], super::IN_DAT_01);
+        let (r, sioe) = do_execute!(["-s", "."], super::IN_DAT_01);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -179,7 +154,7 @@ mod test_s1 {
     #[test]
     fn test_t02() {
         // Child selection
-        let (r, sioe) = do_execute!(&["-s", "\"some\".\"property\""], super::IN_DAT_02);
+        let (r, sioe) = do_execute!(["-s", "\"some\".\"property\""], super::IN_DAT_02);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "\"yay!\"\n",);
         assert!(r.is_ok());
@@ -188,17 +163,17 @@ mod test_s1 {
     #[test]
     fn test_t03() {
         // Index selection
-        let (r, sioe) = do_execute!(&["-s", "\"primes\".[0]"], super::IN_DAT_03);
+        let (r, sioe) = do_execute!(["-s", "\"primes\".[0]"], super::IN_DAT_03);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "7\n",);
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"primes\"[0]"], super::IN_DAT_03);
+        let (r, sioe) = do_execute!(["-s", "\"primes\"[0]"], super::IN_DAT_03);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "7\n",);
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"primes\"[2,0]"], super::IN_DAT_03);
+        let (r, sioe) = do_execute!(["-s", "\"primes\"[2,0]"], super::IN_DAT_03);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "[13,7]\n",);
         assert!(r.is_ok());
@@ -207,7 +182,7 @@ mod test_s1 {
     #[test]
     fn test_t04() {
         // Range selection
-        let (r, sioe) = do_execute!(&["-s", "\"cats\".[1:2]"], super::IN_DAT_04);
+        let (r, sioe) = do_execute!(["-s", "\"cats\".[1:2]"], super::IN_DAT_04);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -215,7 +190,7 @@ mod test_s1 {
         );
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"cats\".[2:1]"], super::IN_DAT_04);
+        let (r, sioe) = do_execute!(["-s", "\"cats\".[2:1]"], super::IN_DAT_04);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -223,7 +198,7 @@ mod test_s1 {
         );
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"cats\".[2:1].[1:0]"], super::IN_DAT_04);
+        let (r, sioe) = do_execute!(["-s", "\"cats\".[2:1].[1:0]"], super::IN_DAT_04);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -231,12 +206,12 @@ mod test_s1 {
         );
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"cats\".[2:1].[0].\"third\""], super::IN_DAT_04);
+        let (r, sioe) = do_execute!(["-s", "\"cats\".[2:1].[0].\"third\""], super::IN_DAT_04);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "\"Misty\"\n",);
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"cats\".[1:]"], super::IN_DAT_04);
+        let (r, sioe) = do_execute!(["-s", "\"cats\".[1:]"], super::IN_DAT_04);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -244,7 +219,7 @@ mod test_s1 {
         );
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"cats\".[:1]"], super::IN_DAT_04);
+        let (r, sioe) = do_execute!(["-s", "\"cats\".[:1]"], super::IN_DAT_04);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -256,12 +231,12 @@ mod test_s1 {
     #[test]
     fn test_t05() {
         // Array selection
-        let (r, sioe) = do_execute!(&["-s", "\"primes\".[]"], super::IN_DAT_05);
+        let (r, sioe) = do_execute!(["-s", "\"primes\".[]"], super::IN_DAT_05);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "[7,11,13]\n",);
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"primes\".[0:2]"], super::IN_DAT_05);
+        let (r, sioe) = do_execute!(["-s", "\"primes\".[0:2]"], super::IN_DAT_05);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "[7,11,13]\n",);
         assert!(r.is_ok());
@@ -270,7 +245,7 @@ mod test_s1 {
     #[test]
     fn test_t06() {
         // Property selection
-        let (r, sioe) = do_execute!(&["-s", "\"object\".{\"a\",\"c\"}"], super::IN_DAT_06);
+        let (r, sioe) = do_execute!(["-s", "\"object\".{\"a\",\"c\"}"], super::IN_DAT_06);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "{\"a\":1,\"c\":3}\n",);
         assert!(r.is_ok());
@@ -279,7 +254,7 @@ mod test_s1 {
     #[test]
     fn test_t07() {
         // Multi-selection
-        let (r, sioe) = do_execute!(&["-s", "\"one\".[2:0],\"two\",\"three\""], super::IN_DAT_07);
+        let (r, sioe) = do_execute!(["-s", "\"one\".[2:0],\"two\",\"three\""], super::IN_DAT_07);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "[[3,2,1],2,3]\n",);
         assert!(r.is_ok());
@@ -288,21 +263,18 @@ mod test_s1 {
     #[test]
     fn test_t08() {
         // Filter
-        let (r, sioe) = do_execute!(&["-s", "\"laptops\"|\"laptop\""], super::IN_DAT_08);
+        let (r, sioe) = do_execute!(["-s", "\"laptops\"|\"laptop\""], super::IN_DAT_08);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "[{\"brand\":\"Apple\",\"options\":[\"a\",\"b\",\"c\"]},{\"brand\":\"Asus\",\"options\":[\"d\",\"e\",\"f\"]}]\n",);
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(
-            &["-s", "\"laptops\"|\"laptop\".\"brand\""],
-            super::IN_DAT_08
-        );
+        let (r, sioe) = do_execute!(["-s", "\"laptops\"|\"laptop\".\"brand\""], super::IN_DAT_08);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "[\"Apple\",\"Asus\"]\n",);
         assert!(r.is_ok());
         //
         let (r, sioe) = do_execute!(
-            &[
+            [
                 "-s",
                 "\"laptops\".[1:0]|\"laptop\".\"brand\",\"laptops\"|\"laptop\".\"brand\""
             ],
@@ -316,7 +288,7 @@ mod test_s1 {
         assert!(r.is_ok());
         //
         let (r, sioe) = do_execute!(
-            &[
+            [
                 "-s",
                 "\"laptops\".[1:0]|\"laptop\"|\"brand\",\"laptops\"|\"laptop\"|\"brand\""
             ],
@@ -333,7 +305,7 @@ mod test_s1 {
     #[test]
     fn test_t09() {
         // Flatten arrays
-        let (r, sioe) = do_execute!(&["-s", "..\"dna\""], super::IN_DAT_09);
+        let (r, sioe) = do_execute!(["-s", "..\"dna\""], super::IN_DAT_09);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -345,12 +317,12 @@ mod test_s1 {
     #[test]
     fn test_t10() {
         // Truncate
-        let (r, sioe) = do_execute!(&["-s", ".!"], super::IN_DAT_10);
+        let (r, sioe) = do_execute!(["-s", ".!"], super::IN_DAT_10);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "{\"foo\":{}}\n",);
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"foo\"!"], super::IN_DAT_10);
+        let (r, sioe) = do_execute!(["-s", "\"foo\"!"], super::IN_DAT_10);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(
             buff!(sioe, sout),
@@ -362,17 +334,17 @@ mod test_s1 {
     #[test]
     fn test_t11() {
         // Special characters
-        let (r, sioe) = do_execute!(&["-s", "\".valid\""], super::IN_DAT_11);
+        let (r, sioe) = do_execute!(["-s", "\".valid\""], super::IN_DAT_11);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "1337\n");
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"\""], super::IN_DAT_11);
+        let (r, sioe) = do_execute!(["-s", "\"\""], super::IN_DAT_11);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "\"yeah!\"\n");
         assert!(r.is_ok());
         //
-        let (r, sioe) = do_execute!(&["-s", "\"\\\"\""], super::IN_DAT_11);
+        let (r, sioe) = do_execute!(["-s", "\"\\\"\""], super::IN_DAT_11);
         assert_eq!(buff!(sioe, serr), "");
         assert_eq!(buff!(sioe, sout), "\"yup, valid too!\"\n");
         assert!(r.is_ok());